+++ src/main/java/com/depromeet/stonebed/domain/fcm/application/FcmService.java
@@ -5,8 +5,6 @@ import com.depromeet.stonebed.domain.fcm.domain.FcmResponseErrorType;
 import com.depromeet.stonebed.domain.fcm.domain.FcmToken;
 import com.depromeet.stonebed.domain.fcm.dto.request.FcmMessageRequest;
 import com.depromeet.stonebed.domain.member.domain.Member;
-import com.depromeet.stonebed.domain.missionRecord.dao.MissionRecordRepository;
-import com.depromeet.stonebed.domain.missionRecord.domain.MissionRecordStatus;
 import com.depromeet.stonebed.global.error.ErrorCode;
 import com.depromeet.stonebed.global.error.exception.CustomException;
 import com.depromeet.stonebed.global.util.MemberUtil;
@@ -16,14 +14,12 @@ import com.google.auth.oauth2.GoogleCredentials;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.*;
 import org.springframework.http.converter.StringHttpMessageConverter;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -36,7 +32,6 @@ public class FcmService {
             "https://fcm.googleapis.com/v1/projects/walwal-dev-fad47/messages:send";
 
     private final FcmRepository fcmRepository;
-    private final MissionRecordRepository missionRecordRepository;
     private final MemberUtil memberUtil;
 
     public void sendMessageToAll(String title, String body) {
@@ -151,57 +146,5 @@ public class FcmService {
 
     public List<String> getAllTokens() {
         return fcmRepository.findAll().stream().map(FcmToken::getToken).toList();
-    }
-
-    public List<String> getIncompleteMissionTokens() {
-        return missionRecordRepository.findAllByStatus(MissionRecordStatus.NOT_COMPLETED).stream()
-                .map(
-                        missionRecord -> {
-                            FcmToken fcmToken =
-                                    fcmRepository
-                                            .findByMember(missionRecord.getMember())
-                                            .orElse(null);
-                            return fcmToken != null ? fcmToken.getToken() : null;
-                        })
-                .filter(token -> token != null)
-                .toList();
-    }
-
-    // 매일 0시 0분에 실행
-    @Scheduled(cron = "0 0 0 * * ?")
-    public void removeInactiveTokens() {
-        LocalDateTime cutoffDate = LocalDateTime.now().minusMonths(2);
-        List<FcmToken> inactiveTokens =
-                fcmRepository.findAll().stream()
-                        .filter(token -> token.getUpdatedAt().isBefore(cutoffDate))
-                        .toList();
-
-        for (FcmToken token : inactiveTokens) {
-            fcmRepository.delete(token);
-        }
-    }
-
-    // 매일 12시 0분에 실행
-    @Scheduled(cron = "0 0 12 * * ?")
-    public void sendDailyNotification() {
-        String title = "정규 메세지 제목!";
-        String body = "정규 메세지 내용!";
-        sendMessageToAll(title, body);
-    }
-
-    // 매일 18시 0분에 실행
-    @Scheduled(cron = "0 0 18 * * ?")
-    public void sendReminderToIncompleteMissions() {
-        List<String> tokens = getIncompleteMissionTokens();
-        String title = "리마인더 메세지 제목!";
-        String body = "리마인더 메세지 내용!";
-
-        for (String token : tokens) {
-            try {
-                sendMessageTo(token, title, body);
-            } catch (IOException e) {
-                log.error("다음 token이 FCM 리마인더 전송에 실패했습니다: {}", token, e);
-            }
-        }
     }
 }
