+++ src/main/java/com/depromeet/stonebed/domain/missionRecord/dao/MissionRecordRepository.java
@@ -4,6 +4,7 @@ import com.depromeet.stonebed.domain.member.domain.Member;
 import com.depromeet.stonebed.domain.mission.domain.MissionHistory;
 import com.depromeet.stonebed.domain.missionRecord.domain.MissionRecord;
 import com.depromeet.stonebed.domain.missionRecord.domain.MissionRecordStatus;
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -13,4 +14,6 @@ public interface MissionRecordRepository
             Member member, MissionHistory missionHistory);
 
     Long countByMemberIdAndStatus(Long memberId, MissionRecordStatus status);
+
+    List<MissionRecord> findAllByStatus(MissionRecordStatus status);
 }
