+++ src/main/java/com/depromeet/stonebed/domain/missionRecord/dao/MissionRecordRepositoryCustom.java
@@ -1,7 +1,6 @@
 package com.depromeet.stonebed.domain.missionRecord.dao;
 
 import com.depromeet.stonebed.domain.missionRecord.domain.MissionRecord;
-import com.depromeet.stonebed.domain.missionRecord.domain.MissionRecordStatus;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
@@ -11,6 +10,4 @@ public interface MissionRecordRepositoryCustom {
 
     List<MissionRecord> findByMemberIdAndCreatedAtFromWithPagination(
             Long memberId, LocalDateTime createdAt, Pageable pageable);
-
-    List<MissionRecord> findAllByStatus(MissionRecordStatus status);
 }
